PiratePICProg: Commited patches from AndThen (PIC32MX), PIC32 still needs work. Added verbose debugging print (so far buspirate and serial module). Fixed PIC24 readout support, fixed endiannes of read data (BP_PIC424Read), added correct unpacking of read data.

+++ PiratePICprog/software/pic16.c
@@ -12,7 +12,7 @@ static uint32_t PIC16_ExitICSP(struct picprog_t *p, enum icsp_t type)
 	return 1;
 }
 
-static uint32_t PIC16_ReadID(struct picprog_t *p, uint16_t *id, uint16_t *rev)
+static uint32_t PIC16_ReadID(struct picprog_t *p, uint32_t *id, uint16_t *rev)
 {
 	return 1;
 }
